update user report profile
